Saldırı animasyonları için blendtree eklendi

+++ Assets/Scripts/Characters/Enemy/EnemyTypes/Mushroom.cs
@@ -68,7 +68,7 @@ public class Mushroom : Enemy
         if (currentAttackRate <= 0 && canAttack)
         {
             animator.SetTrigger(EnemyAnimationParametres.attack);
-            animator.SetInteger(EnemyAnimationParametres.attackIndex, Random.Range(0, 2));
+            animator.SetFloat(EnemyAnimationParametres.attackIndex, Random.Range(0, 2));
             canAttack = false;
         }
     }
